% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/external_functions.R
\name{fair_expand}
\alias{fair_expand}
\title{Expand Json Files}
\usage{
fair_expand(json_file)
}
\arguments{
\item{json_file}{Json-ld created from the fairify_data function}
}
\value{
Expanded json-ld file
}
\description{
Expand Json Files
}
\examples{
json_file <- '{
"@context": "http://schema.org/",
"@type": "Person",
"name": "Jane Doe",
"jobTitle": "Professor",
"telephone": "(425) 123-4567",
"url": "http://www.janedoe.com"}'

fair_expand(json_file)
}
