% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/powerplant_fair.R
\name{powerplant_fair}
\alias{powerplant_fair}
\title{Powerplant FAIR}
\usage{
powerplant_fair(specific_data)
}
\arguments{
\item{specific_data}{This is a dataframe with specially formatted metadata}
}
\value{
Returns a .json-ld file with desired metadata
}
\description{
Given a table of metadata, this function generates a .json-ld file with the metadata for a solar powerplant
}
\examples{
powerplant_fair(test_metadata)

}
