% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QUV1_fair.R
\name{QUV1_fair}
\alias{QUV1_fair}
\title{QUV1 FAIR}
\usage{
QUV1_fair(method_in)
}
\arguments{
\item{method_in}{The method that was used for the experiment run}
}
\value{
Returns a .json-ld file with desired metadata
}
\description{
Automatically generates a .json-ld file with metadata for a given run of the QUV1 instrument
}
\examples{
QUV1_fair('hot-quv')

}
