% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fairify_data.R
\name{rename_fair}
\alias{rename_fair}
\title{Rename Column Names}
\usage{
rename_fair(dataframe, domain)
}
\arguments{
\item{dataframe}{User input dataframe}

\item{domain}{Json-LD template to be filled by function}
}
\value{
Dataframe with FAIRified column names
}
\description{
Rename Column Names
}
