% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fairify_data.R
\docType{data}
\name{simulationXRayDiffraction}
\alias{simulationXRayDiffraction}
\title{This is data to be included in the package}
\description{
This is data to be included in the package
}
\keyword{internal}
