% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sunsvoc_fair.R
\name{sunsvoc_fair}
\alias{sunsvoc_fair}
\title{SunsVoc FAIR}
\usage{
sunsvoc_fair(method_in)
}
\arguments{
\item{method_in}{The method that was used for the experiment run}
}
\value{
Returns a .json-ld file with desired metadata
}
\description{
Automatically generates a .json-ld file with metadata for a given run of the SunsVoc instrument
}
\examples{
sunsvoc_fair('default.svi')

}
