% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fairify_data.R
\name{json_conversion}
\alias{json_conversion}
\title{Convert Dataframe to Json from a Template}
\usage{
json_conversion(row, unit_dataframe, domain, i, data_list)
}
\arguments{
\item{row}{Row of a user input csv that includes metadata}

\item{unit_dataframe}{User input units}

\item{domain}{Json-LD template to be filled by function}

\item{i}{Iteration of the loop tracker}

\item{data_list}{List of columns included in any of your raw dataframes \cr}
}
\value{
FAIRified .json-ld text document
}
\description{
Convert Dataframe to Json from a Template
}
